+++ model_config.yaml
@@ -2,7 +2,7 @@ grid_config:
     sklearn.ensemble.RandomForestClassifier:
         n_estimators: [100, 500, 1000]
         criterion: [gini]
-        max_depth: [10, 20, 30, 50]
+        max_depth: [5, 10, 20, 30, 50]
         max_features: [sqrt, log2, 1.0]
         min_samples_split: [5, 10, 20]
         n_jobs: [2]
